Reorganização dos pacotes em [Application, Entidade, Enums, Interfaces e Repositório].
Limpeza de arquivos não utilizados
