+++ src/application/Mensagem.java
@@ -1,4 +1,4 @@
-package agendaMain;
+package application;
 
 public enum Mensagem {
 	 AVISO,ERRO,PADRAO;
