+++ src/entidade/Calendario.java
@@ -1,4 +1,4 @@
-package classesAux;
+package entidade;
 
 import java.util.Calendar;
 import java.util.GregorianCalendar;
