+++ src/entidade/Contato.java
@@ -1,4 +1,4 @@
-package classesAux;
+package entidade;
 
 public class Contato {
 	private String nome;
