+++ src/entidade/Evento.java
@@ -1,13 +1,9 @@
-package compromissos;
-import classesAux.Calendario;
-import classesAux.Endereco;
-import classesAux.Contato;
-public class Reuniao extends CompromissoGeral {
+package entidade;
+
+public class Evento extends CompromissoGeral{
 	Calendario dataHoraFim;
 	Endereco local;
-	//Contato[] participantesObrigatorios;
 	Contato[] participantes;
-	Contato coordenador;
 	public Calendario getDataHoraFim() {
 		return dataHoraFim;
 	}
@@ -26,12 +22,5 @@ public class Reuniao extends CompromissoGeral {
 	public void setParticipantes(Contato[] participantes) {
 		this.participantes = participantes;
 	}
-	public Contato getCoordenador() {
-		return coordenador;
-	}
-	public void setCoordenador(Contato coordenador) {
-		this.coordenador = coordenador;
-	}
-	
 	
 }
