+++ src/entidade/Lembrete.java
@@ -1,4 +1,5 @@
-package compromissos;
+package entidade;
+
 public class Lembrete extends CompromissoGeral{
 	
 }
