+++ src/enums/TipoAgendamento.java
@@ -1,4 +1,4 @@
-package regras;
+package enums;
 
 public enum TipoAgendamento {
 	REUNIAO, LEMBRENTE, EVENTO, OUTRO;
