+++ src/interfaces/InterfaceRepositorio.java
@@ -1,5 +1,8 @@
-package compromissos;
-public interface interfaceReuniao {
+package interfaces;
+
+import entidade.Reuniao;
+
+public interface InterfaceRepositorio {
 		public boolean inserirReuniao(Reuniao reuniao);
 		public boolean excluirReuniao(Reuniao reuniao);
 		public Reuniao consultarReuniao(Reuniao reuniao);
